fixed a problem with architecture, added mappers
