+++ src/main/java/ru/monetarys/config/RabbitMqConfiguration.java
@@ -12,7 +12,7 @@ import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
 import org.springframework.amqp.support.converter.MessageConverter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import ru.monetarys.services.clientprofile.ApplicationProperties;
+import ru.monetarys.integration.ApplicationProperties;
 
 @Configuration
 @RequiredArgsConstructor
