+++ src/main/java/ru/monetarys/integration/domain/ClientContactsInfo.java
@@ -1,0 +1,10 @@
+package ru.monetarys.integration.domain;
+
+import lombok.Data;
+
+@Data
+public class ClientContactsInfo {
+
+    private String phoneNumber;
+
+}
