+++ src/main/java/ru/monetarys/integration/domain/ClientGeneralInfo.java
@@ -1,0 +1,16 @@
+package ru.monetarys.integration.domain;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class ClientGeneralInfo {
+
+    private String guid;
+
+    private ClientPersonalInfo personalInfo;
+    private List<ClientAccountInfo> accountList;
+    private ClientContactsInfo contacts;
+
+}
