+++ src/main/java/ru/monetarys/integration/domain/ClientPersonalInfo.java
@@ -1,0 +1,13 @@
+package ru.monetarys.integration.domain;
+
+import lombok.Data;
+
+@Data
+public class ClientPersonalInfo {
+
+    private String lastName;
+    private String firstName;
+    private String citizenshipCountryCode;
+    private String middleName;
+
+}
