+++ src/main/java/ru/monetarys/integration/rs/ClientPersonalInfoRs.java
@@ -1,0 +1,18 @@
+package ru.monetarys.integration.rs;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+
+@Data
+public class ClientPersonalInfoRs {
+
+    @JsonProperty(value = "lastName", required = true)
+    private String lastName;
+
+    @JsonProperty(value = "firstName", required = true)
+    private String firstName;
+
+    private String citizenshipCountryCode;
+    private String middleName;
+
+}
