+++ src/main/java/ru/monetarys/integration/service/ClientProfileService.java
@@ -1,0 +1,9 @@
+package ru.monetarys.integration.service;
+
+import ru.monetarys.integration.domain.ClientGeneralInfo;
+
+public interface ClientProfileService {
+
+    ClientGeneralInfo getClientInfoByGUID(String guid);
+
+}
