+++ src/main/java/ru/monetarys/logic/impl/TransferManager.java
@@ -1,0 +1,16 @@
+package ru.monetarys.logic.impl;
+
+import ru.monetarys.integration.domain.ClientAccountInfo;
+import ru.monetarys.integration.domain.ClientGeneralInfo;
+import ru.monetarys.web.domain.TransferRequest;
+import ru.monetarys.web.ro.TransferResponseRo;
+
+public interface TransferManager {
+
+    TransferResponseRo transferMoney(TransferRequest transferRequest);
+
+    void validatePayer(TransferRequest transferRequest, ClientAccountInfo payerAccount);
+
+    void validatePayee(ClientGeneralInfo payee, ClientGeneralInfo payer, ClientAccountInfo payeeAccount, ClientAccountInfo payerAccount);
+
+}
