--- src/main/java/ru/monetarys/services/clientprofile/ClientProfileService.java
@@ -1,9 +1,0 @@
-package ru.monetarys.services.clientprofile;
-
-import ru.monetarys.dto.ClientGeneralInfo;
-
-public interface ClientProfileService {
-
-    ClientGeneralInfo getClientInfoByGUID(String guid);
-
-}
