--- src/main/java/ru/monetarys/transfer/entities/Payee.java
@@ -1,14 +1,0 @@
-package ru.monetarys.transfer.entities;
-
-import lombok.Data;
-
-@Data
-public class Payee {
-
-    private String guid;
-    private String accountId;
-    private String account;
-    private String cardNumber;
-    private String phone;
-
-}
