--- src/main/java/ru/monetarys/transfer/entities/PayeeAccount.java
@@ -1,12 +1,0 @@
-package ru.monetarys.transfer.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class PayeeAccount {
-
-    private String account;
-
-}
