--- src/main/java/ru/monetarys/transfer/entities/Payer.java
@@ -1,11 +1,0 @@
-package ru.monetarys.transfer.entities;
-
-import lombok.Data;
-
-@Data
-public class Payer {
-
-    private String guid;
-    private String account;
-
-}
