--- src/main/java/ru/monetarys/transfer/entities/PayerAccount.java
@@ -1,12 +1,0 @@
-package ru.monetarys.transfer.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class PayerAccount {
-
-    private String account;
-
-}
