--- src/main/java/ru/monetarys/transfer/entities/Transaction.java
@@ -1,13 +1,0 @@
-package ru.monetarys.transfer.entities;
-
-import lombok.Data;
-
-import java.math.BigDecimal;
-
-@Data
-public class Transaction {
-
-    private BigDecimal amount;
-    private String messageToPayee;
-
-}
