--- src/main/java/ru/monetarys/transfer/entities/TransferRequest.java
@@ -1,12 +1,0 @@
-package ru.monetarys.transfer.entities;
-
-import lombok.Data;
-
-@Data
-public class TransferRequest {
-
-    private Payer payer;
-    private Payee payee;
-    private Transaction transaction;
-
-}
