--- src/main/java/ru/monetarys/transfer/entities/TransferResponse.java
@@ -1,14 +1,0 @@
-package ru.monetarys.transfer.entities;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class TransferResponse {
-
-    private TransactionResponse transaction;
-    private PayerAccount payerAccount;
-    private PayeeAccount payeeAccount;
-
-}
