+++ src/main/java/ru/monetarys/web/domain/Payee.java
@@ -1,0 +1,14 @@
+package ru.monetarys.web.domain;
+
+import lombok.Data;
+
+@Data
+public class Payee {
+
+    private String guid;
+    private String accountId;
+    private String account;
+    private String cardNumber;
+    private String phone;
+
+}
