+++ src/main/java/ru/monetarys/web/domain/Payer.java
@@ -1,0 +1,11 @@
+package ru.monetarys.web.domain;
+
+import lombok.Data;
+
+@Data
+public class Payer {
+
+    private String guid;
+    private String account;
+
+}
