+++ src/main/java/ru/monetarys/web/domain/TransferRequest.java
@@ -1,0 +1,12 @@
+package ru.monetarys.web.domain;
+
+import lombok.Data;
+
+@Data
+public class TransferRequest {
+
+    private Payer payer;
+    private Payee payee;
+    private Transaction transaction;
+
+}
