+++ src/main/java/ru/monetarys/web/ro/PayeeRo.java
@@ -1,0 +1,14 @@
+package ru.monetarys.web.ro;
+
+import lombok.Data;
+
+@Data
+public class PayeeRo {
+
+    private String guid;
+    private String accountId;
+    private String account;
+    private String cardNumber;
+    private String phone;
+
+}
