+++ src/main/java/ru/monetarys/web/ro/PayerAccountResponseRo.java
@@ -1,0 +1,13 @@
+package ru.monetarys.web.ro;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class PayerAccountResponseRo {
+
+    private String account;
+
+}
