+++ src/main/java/ru/monetarys/web/ro/PayerRo.java
@@ -1,0 +1,11 @@
+package ru.monetarys.web.ro;
+
+import lombok.Data;
+
+@Data
+public class PayerRo {
+
+    private String guid;
+    private String account;
+
+}
