+++ src/main/java/ru/monetarys/web/ro/TransactionRo.java
@@ -1,0 +1,13 @@
+package ru.monetarys.web.ro;
+
+import lombok.Data;
+
+import java.math.BigDecimal;
+
+@Data
+public class TransactionRo {
+
+    private BigDecimal amount;
+    private String messageToPayee;
+
+}
