+++ src/main/java/ru/monetarys/web/ro/TransferRequestRo.java
@@ -1,0 +1,12 @@
+package ru.monetarys.web.ro;
+
+import lombok.Data;
+
+@Data
+public class TransferRequestRo {
+
+    private PayerRo payer;
+    private PayeeRo payee;
+    private TransactionRo transaction;
+
+}
