+++ src/main/java/ru/monetarys/web/service/TransferService.java
@@ -1,0 +1,10 @@
+package ru.monetarys.web.service;
+
+import ru.monetarys.web.ro.TransferRequestRo;
+import ru.monetarys.web.ro.TransferResponseRo;
+
+public interface TransferService {
+
+    TransferResponseRo transferMoney(TransferRequestRo transferRequestRo);
+
+}
